[EFL] EWebLauncher is executed as full screen with device pixel ratio https://bugs.webkit.org/show_bug.cgi?id=120282
Reviewed by Gyuyoung Kim.

* EWebLauncher/main.c:
Use double instead of float for device_pixel_ratio which is passed to ECORE_GETOPT_VALUE_DOUBLE.


Canonical link: https://commits.webkit.org/138241@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@154587 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/EWebLauncher/main.c
@@ -150,7 +150,7 @@ typedef struct _User_Arguments {
     char *engine;
     Eina_Bool quitOption;
     char *backingStore;
-    float device_pixel_ratio;
+    double device_pixel_ratio;
     Eina_Bool enableEncodingDetector;
     Eina_Bool enableTiledBackingStore;
     Eina_Bool isFlattening;
